fix: GenerateMetadataProps Readonly 추가
