+++ src/components/groupDetail/GroupAlbum.tsx
@@ -5,28 +5,22 @@ import GroupInfoBox from './GroupInfoBox';
 import Loading from '@/app/loading';
 import useIntersect from '@/hooks/byUse/useIntersection';
 import { useIsOpen } from '@/hooks/byUse/useIsOpen';
-import useMediaQuery from '@/hooks/byUse/useMediaQuery';
 import { getGroupPostsImagesQuery } from '@/hooks/queries/post/useGroupPostsQuery';
 import { Modal } from '@/stories/Modal';
 import { type GroupDetailMode, type GroupInfo } from '@/types/groupTypes';
 import Link from 'next/link';
-import { useEffect, useState } from 'react';
 
 type Props = {
   groupId: string;
   groupInfo: GroupInfo;
   setMode: React.Dispatch<React.SetStateAction<GroupDetailMode>>;
+  desktop: boolean;
 };
 
-const GroupAlbum = ({ groupId, groupInfo, setMode }: Props) => {
+const GroupAlbum = ({ groupId, groupInfo, setMode, desktop }: Props) => {
   const { data, isPending, isError, error, hasNextPage, fetchNextPage, isFetchingNextPage, isFetching } =
     getGroupPostsImagesQuery(groupId);
   const [updateModal, handleUpdateModal] = useIsOpen();
-  const isDesktop = useMediaQuery('(min-width: 1200px)');
-  const [desktop, setDesktop] = useState(false);
-  useEffect(() => {
-    setDesktop(isDesktop);
-  }, [isDesktop]);
 
   //NOTE - 고장
   const observerRef = useIntersect(async (entry, observer) => {
